requirements.txt
langchain_openai 
langchain_core
python-dotenv
langchain_community
langchain_ollama
langchain_google_genai
langserve

chromadb
faiss-cpu

streamlit
ipykernel
fastapi
uvicorn
pypdf
bs4
PyPDF2